fix compatibility with opensmtpd 5.9+
This mirrors changes done in [1].

When a table from extras is used (tested with passwd and postgres),
opensmtpd failes to start (on Alpine Linux):

    passwd[7508]: warn: table-proc: bogus data
    passwd[7508]: fatal: table-proc: exiting
    warn: table-proc: pipe closed
    fatal: table-proc: exiting

The error is printed by the table process and it's cased by mismatch
between expected and actual size of the struct table_open_params.

Note: I've originally tried to port even changes of includes from [1],
but that would need to add many new includes to most of the extras.
Thus I've eventually decided that it's probably not the right approach.

Fixes: https://github.com/OpenSMTPD/OpenSMTPD/issues/816

[1]: https://github.com/OpenSMTPD/OpenSMTPD/commit/5dfecad33e1301343473f7e9a6e425cdd11b9c3f