Bind with LDAP user credentials to authenticate user
If enabled using "credentials_bind" configuration option, the table will
find the "dn" of the user in the LDAP directory (using "credentials_filter"
option) and then try to log into the LDAP server using this dn and the
password of the user.

This is useful for servers like OpenLDAP which doesn't store the password
using the crypt method. We can't just query the user and password with
K_CREDENTIALS table service because crypt_checkpass(3) will fail. So use
the new K_CHECKUSER instead and let the backend verify the password for us.

E.g. in your /etc/mail/ldapd.conf:
...
credentials_filter      (&(objectClass=inetOrgPerson)(cn=%s))
credentials_bind
...