table-ldap: Strip {CRYPT} prefix from userPassword
Both OpenBSD's ldapd and OpenLDAP use RFC2307 passwords for the
userPassword field; these are either plaintext, or in the form of
"{scheme}hash". A scheme of CRYPT specifies the hash can be checked
using crypt(3). On OpenBSD at least, this will be in the same format as
OpenSMTPD's expected hash.

If the credentials lookup uses the well-known attribute userPassword to
retrieve the password hash, and the result is prefixed with {CRYPT},
remove this prefix to allow direct use of userPassword.